requirements.txt
opencv-python
numpy
mss
mysql-connector-python
pandas
scipy
Pillow
pyautogui